feat: add support for new set imei endpoint
