+++ src/Model/Order/OrderItem.php
@@ -196,6 +196,11 @@ class OrderItem implements JsonSerializable
     protected $salesType;
 
     /**
+     * @var string|null
+     */
+    protected $imei;
+
+    /**
      * @param mixed[]|null $extraAttributes
      */
     final public static function fromOrderItem(
@@ -235,7 +240,8 @@ class OrderItem implements JsonSerializable
         ?DateTimeImmutable $createdAt,
         ?DateTimeImmutable $updatedAt,
         ?string $returnStatus,
-        ?string $salesType = null
+        ?string $salesType = null,
+        ?string $imei = null
     ): OrderItem {
         $orderItem = new self();
 
@@ -276,6 +282,7 @@ class OrderItem implements JsonSerializable
         $orderItem->updatedAt = $updatedAt;
         $orderItem->returnStatus = $returnStatus;
         $orderItem->salesType = $salesType;
+        $orderItem->imei = $imei;
 
         return $orderItem;
     }
@@ -484,6 +491,16 @@ class OrderItem implements JsonSerializable
         return !empty($this->salesType) ? $this->salesType : null;
     }
 
+    public function getImei(): ?string
+    {
+        return !empty($this->imei) ? $this->imei : null;
+    }
+
+    public function setImei(?string $imei): void
+    {
+        $this->imei = $imei;
+    }
+
     public function jsonSerialize(): stdClass
     {
         $serialized = new stdClass();
@@ -524,6 +541,7 @@ class OrderItem implements JsonSerializable
         $serialized->updatedAt = $this->updatedAt;
         $serialized->returnStatus = $this->returnStatus;
         $serialized->salesType = $this->salesType;
+        $serialized->imei = $this->imei;
 
         return $serialized;
     }
