+++ src/Service/GlobalOrderManager.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Linio\SellerCenter\Service;
+
+class GlobalOrderManager extends BaseOrderManager
+{
+}
