+++ tests/Unit/Order/OrderItemsTest.php
@@ -8,6 +8,7 @@ use Linio\SellerCenter\Factory\Xml\Order\OrderItemsFactory;
 use Linio\SellerCenter\LinioTestCase;
 use Linio\SellerCenter\Model\Order\OrderItem;
 use Linio\SellerCenter\Model\Order\OrderItems;
+use Linio\SellerCenter\Transformer\Order\OrderItemsTransformer;
 
 class OrderItemsTest extends LinioTestCase
 {
@@ -41,5 +42,20 @@ class OrderItemsTest extends LinioTestCase
 
         $this->assertInstanceOf(OrderItems::class, $orderItems);
         $this->assertContainsOnlyInstancesOf(OrderItem::class, $orderItems->all());
+    }
+
+    public function testTransformsAOrderItemsObjectIntoAnXmlRepresentation(): void
+    {
+        $simpleXml = simplexml_load_string($this->getSchema('Order/OrderItemsResponse.xml'));
+
+        $orderItems = OrderItemsFactory::make($simpleXml->Body);
+
+        $orderItemsXml = OrderItemsTransformer::orderItemsImeiAsXml($orderItems);
+
+        $orderItems = $orderItems->all();
+        $orderItem = !empty($orderItems) ? reset($orderItems) : [];
+
+        $this->assertEquals($orderItem->getOrderItemId(), (int) $orderItemsXml->OrderItem->OrderItemId);
+        $this->assertEquals($orderItem->getImei(), $orderItemsXml->OrderItem->Imei);
     }
 }
