+++ tests/_schemas/Order/OrderItem.xml
@@ -37,4 +37,6 @@
   <UpdatedAt>2018-10-14 13:30:50</UpdatedAt>
   <ReturnStatus>%s</ReturnStatus>
   <SalesType>PURE_MARKETPLACE</SalesType>
+  <SalesType>PURE_MARKETPLACE</SalesType>
+  <Imei>qwe12eqw165qw4-qw7e89q654</Imei>
 </OrderItem>
