+++ tests/_schemas/Order/OrderItemsResponse.xml
@@ -45,6 +45,7 @@
         <CreatedAt>2018-10-13 23:08:34</CreatedAt>
         <UpdatedAt>2018-10-14 13:30:50</UpdatedAt>
         <ReturnStatus/>
+        <Imei>weqweqwe-123123</Imei>
       </OrderItem>
     </OrderItems>
   </Body>
