+++ tests/_schemas/Order/OrderWithOrderItems.json
@@ -58,7 +58,8 @@
       "createdAt": null,
       "updatedAt": null,
       "returnStatus": null,
-      "salesType": null
+      "salesType": null,
+      "imei": null
     }
   ]
 }
