+++ tests/_schemas/Order/SetImeiRequest.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Request>
+  <OrderItem>
+    <OrderItemId>6750999</OrderItemId>
+    <Imei>qweqweasd-qw213123</Imei>
+  </OrderItem>
+</Request>
